requirements.txt
bottle
tellcore-py
